FGZ and No-Tick rules + Refactoring
Now the Stones structure now keeps array + flags which stones are
in play and which belongs to which team. And it keeps positions
only - the velocity and acceleration are in MovingStone which is
used by delivery::Process now.

The process was also changed: now every stone has separate time
quantums, which durations are computed in a clever way.

The free guards and center guard list is remembered at the turn
beggining, and violations are checked after delivery finish.
If there is some violation, the turn is not progressed
automatically and a message is displayed.

There is a new module "simulation" taking over turn submodules.